Moved histogram agg tests to correct test package
