+++ x-pack/plugin/analytics/src/test/java/org/elasticsearch/xpack/analytics/aggregations/metrics/HDRPreAggregatedPercentileRanksAggregatorTests.java
@@ -3,10 +3,10 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-package org.elasticsearch.xpack.analytics.mapper;
+package org.elasticsearch.xpack.analytics.aggregations.metrics;
 
-import com.tdunning.math.stats.Centroid;
-import com.tdunning.math.stats.TDigest;
+import org.HdrHistogram.DoubleHistogram;
+import org.HdrHistogram.DoubleHistogramIterationValue;
 import org.apache.lucene.document.BinaryDocValuesField;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.index.IndexReader;
@@ -19,27 +19,25 @@ import org.elasticsearch.index.mapper.MappedFieldType;
 import org.elasticsearch.plugins.SearchPlugin;
 import org.elasticsearch.search.aggregations.AggregationBuilder;
 import org.elasticsearch.search.aggregations.AggregatorTestCase;
-import org.elasticsearch.search.aggregations.metrics.InternalTDigestPercentileRanks;
+import org.elasticsearch.search.aggregations.metrics.InternalHDRPercentileRanks;
 import org.elasticsearch.search.aggregations.metrics.Percentile;
 import org.elasticsearch.search.aggregations.metrics.PercentileRanks;
 import org.elasticsearch.search.aggregations.metrics.PercentileRanksAggregationBuilder;
 import org.elasticsearch.search.aggregations.metrics.PercentilesConfig;
 import org.elasticsearch.search.aggregations.metrics.PercentilesMethod;
-import org.elasticsearch.search.aggregations.metrics.TDigestState;
 import org.elasticsearch.search.aggregations.support.AggregationInspectionHelper;
 import org.elasticsearch.search.aggregations.support.CoreValuesSourceType;
 import org.elasticsearch.search.aggregations.support.ValuesSourceType;
 import org.elasticsearch.xpack.analytics.AnalyticsPlugin;
 import org.elasticsearch.xpack.analytics.aggregations.support.AnalyticsValuesSourceType;
+import org.elasticsearch.xpack.analytics.mapper.HistogramFieldMapper;
 import org.hamcrest.Matchers;
 
 import java.io.IOException;
-import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
 
-
-public class TDigestPreAggregatedPercentileRanksAggregatorTests extends AggregatorTestCase {
+public class HDRPreAggregatedPercentileRanksAggregatorTests extends AggregatorTestCase {
 
     @Override
     protected List<SearchPlugin> getSearchPlugins() {
@@ -48,9 +46,9 @@ public class TDigestPreAggregatedPercentileRanksAggregatorTests extends Aggregat
 
     @Override
     protected AggregationBuilder createAggBuilderForTypeTest(MappedFieldType fieldType, String fieldName) {
-        return new PercentileRanksAggregationBuilder("tdigest_percentiles", new double[]{1.0})
+        return new PercentileRanksAggregationBuilder("hdr_percentiles", new double[]{1.0})
             .field(fieldName)
-            .percentilesConfig(new PercentilesConfig.TDigest());
+            .percentilesConfig(new PercentilesConfig.Hdr());
     }
 
     @Override
@@ -63,18 +61,19 @@ public class TDigestPreAggregatedPercentileRanksAggregatorTests extends Aggregat
     }
 
     private BinaryDocValuesField getDocValue(String fieldName, double[] values) throws IOException {
-        TDigest histogram = new TDigestState(100.0); //default
+        DoubleHistogram histogram = new DoubleHistogram(3);//default
         for (double value : values) {
-            histogram.add(value);
+            histogram.recordValue(value);
         }
         BytesStreamOutput streamOutput = new BytesStreamOutput();
-        histogram.compress();
-        Collection<Centroid> centroids = histogram.centroids();
-        Iterator<Centroid> iterator = centroids.iterator();
-        while ( iterator.hasNext()) {
-            Centroid centroid = iterator.next();
-            streamOutput.writeVInt(centroid.count());
-            streamOutput.writeDouble(centroid.mean());
+        DoubleHistogram.RecordedValues recordedValues = histogram.recordedValues();
+        Iterator<DoubleHistogramIterationValue> iterator = recordedValues.iterator();
+        while (iterator.hasNext()) {
+            DoubleHistogramIterationValue value = iterator.next();
+            long count = value.getCountAtValueIteratedTo();
+            streamOutput.writeVInt(Math.toIntExact(count));
+            double d = value.getValueIteratedTo();
+            streamOutput.writeDouble(d);
         }
         return new BinaryDocValuesField(fieldName, streamOutput.bytes().toBytesRef());
     }
@@ -86,31 +85,27 @@ public class TDigestPreAggregatedPercentileRanksAggregatorTests extends Aggregat
             doc.add(getDocValue("field", new double[] {3, 0.2, 10}));
             w.addDocument(doc);
 
-            PercentileRanksAggregationBuilder aggBuilder = new PercentileRanksAggregationBuilder("my_agg", new double[] {0.1, 0.5, 12})
+            PercentileRanksAggregationBuilder aggBuilder = new PercentileRanksAggregationBuilder("my_agg", new double[]{0.1, 0.5, 12})
                     .field("field")
-                    .method(PercentilesMethod.TDIGEST);
-            MappedFieldType fieldType = new HistogramFieldMapper.Builder("number").fieldType();
+                    .method(PercentilesMethod.HDR);
+            MappedFieldType fieldType = new HistogramFieldMapper.Builder("field").fieldType();
             fieldType.setName("field");
             try (IndexReader reader = w.getReader()) {
                 IndexSearcher searcher = new IndexSearcher(reader);
-                PercentileRanks ranks = search(searcher, new MatchAllDocsQuery(), aggBuilder, fieldType);
+                PercentileRanks ranks = searchAndReduce(searcher, new MatchAllDocsQuery(), aggBuilder, fieldType);
                 Iterator<Percentile> rankIterator = ranks.iterator();
                 Percentile rank = rankIterator.next();
                 assertEquals(0.1, rank.getValue(), 0d);
-                // TODO: Fix T-Digest: this assertion should pass but we currently get ~15
-                // https://github.com/elastic/elasticsearch/issues/14851
-                // assertThat(rank.getPercent(), Matchers.equalTo(0d));
+                assertThat(rank.getPercent(), Matchers.equalTo(0d));
                 rank = rankIterator.next();
                 assertEquals(0.5, rank.getValue(), 0d);
                 assertThat(rank.getPercent(), Matchers.greaterThan(0d));
                 assertThat(rank.getPercent(), Matchers.lessThan(100d));
                 rank = rankIterator.next();
                 assertEquals(12, rank.getValue(), 0d);
-                // TODO: Fix T-Digest: this assertion should pass but we currently get ~59
-                // https://github.com/elastic/elasticsearch/issues/14851
-                // assertThat(rank.getPercent(), Matchers.equalTo(100d));
+                assertThat(rank.getPercent(), Matchers.equalTo(100d));
                 assertFalse(rankIterator.hasNext());
-                assertTrue(AggregationInspectionHelper.hasValue(((InternalTDigestPercentileRanks)ranks)));
+                assertTrue(AggregationInspectionHelper.hasValue((InternalHDRPercentileRanks)ranks));
             }
         }
     }
